+++ src/pages/Quote.js
